Se modifico la clase principal con la finalidad de que el usuario pueda ingresar los datos para cada unos de los ejes, asi como también, muestra si ambos rectangulos estan uno sobre otro, estan juntos o separados

+++ Laboratorio1/Ejercicio1/Coordenada.java
@@ -8,7 +8,7 @@ public class Coordenada {
         this(0,0);
     }
 
-    public Coordenada(int x, int y) {
+    public Coordenada(double x, double y) {
         this.x = x;
         this.y = y;
     }
